// src/Model/productItemsData.model.ts
export interface IAnalyticsData {
  SaleTotalTransactions: number;
  PurchaseTotalTransactions?: number;
  ItemData?: number;
  ItemCategoryData?: number;
  ExpenseData?: number;
  PartyData?: number;
  LastSaleDoneOn?: number;
  LastPurchaseDoneOn?: number;
  LastItemAddedOn?: number;
  LastCategoryCreatedOn?: number;
  LastExpenseCreatedOn?: number;
  LastAppUsedOn?: number;
  TotalTimeSpent?: number;
  LastAppStartedon?: number;
  MobileNumber?: number;
  TimeSpent?: number;
  DailyAccountsReport?: number;
  GSTR1Report?: number;
  GSTR2Report?: number;
  GSTR3BReport?: number;
  InventoryStockReport?: number;
  LowStockSummaryReport?: number;
  PartyStatementReport?: number;
  ProductLevelInventoryReport?: number;
  ProfitAndLostReport?: number;
  PurchaseDraftReport?: number;
  PurchaseReport?: number;
  SaleReport?: number;
  SalesDraftReport?: number;
  AppsubscriptionPage?: number;
  subscriptionContactInformation?: number;
  MonthlySubscriptionSuccess?: number;
  AnuallySubscriptionSucess?: number;
  FailurePayment?: number;
  YearEndingReport?: number;
  InsertedDate?: string;
  SalesAdded?: string;
  PurchasesAdded?: string;
  ExpensesAdded?: string;
  PartiesAdded?: string;
  ItemsAdded?: string;
  CategoriesAdded?: string;
  ID?: number;
  CampaignID?: number;
  SaleAppWalkThrough?: string;
  InventoryAppWalkThrough?: string;
}